requirements.txt
boto3
django
django-storages
django-recaptcha
django-debug-toolbar
django-ses[events]
django-sesame
django-widget-tweaks
django-parler
django-ninja
pgvector
psycopg2-binary
gunicorn
uvicorn[standard]
python-slugify
openai
openai-agents
python-dotenv
python-dateutil
feedparser
redis
hiredis
django-redis
markdown
google-api-python-client
youtube-transcript-api
yt-dlp
pypalettes
whitenoise
celery
django-celery-results
pillow
whitenoise
